+++ protostar/commands/test/starkware/cheatable_syscall_handler.py
@@ -52,32 +52,6 @@ class CheatableSysCallHandler(BusinessLogicSysCallHandler):
             ]
         return super()._get_block_timestamp()
 
-    def set_caller_address(
-        self, addr: int, target_contract_address: Optional[int] = None
-    ):
-        target = (
-            target_contract_address
-            if target_contract_address
-            else self.contract_address
-        )
-        if target in self.cheatable_state.pranked_contracts_map:
-            raise CheatableSysCallHandlerException(
-                f"Contract with address {target} has been already pranked"
-            )
-        self.cheatable_state.pranked_contracts_map[target] = addr
-
-    def reset_caller_address(self, target_contract_address: Optional[int] = None):
-        target = (
-            target_contract_address
-            if target_contract_address
-            else self.contract_address
-        )
-        if target not in self.cheatable_state.pranked_contracts_map:
-            raise CheatableSysCallHandlerException(
-                f"Contract with address {target} has not been pranked"
-            )
-        del self.cheatable_state.pranked_contracts_map[target]
-
     def _get_caller_address(
         self,
         segments: MemorySegmentManager,
